Fix unox.rb formula
There seems to be a bug in Homebrew where it incorrectly guesses the extension based on the URL when creating a temporary file. This causes `gunzip` to try unpacking a file without extension which fails silently.

This PR fixes #6 by replacing URL without an extension to one with an extension.